fix package directives, add session cookie to protected route responses
