+++ bookmarket-backend/src/main/java/edu/oakland/sophomoreproject/authorization/CookieExtractor.java
@@ -1,4 +1,4 @@
-package authorization;
+package edu.oakland.sophomoreproject.authorization;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
