+++ bookmarket-backend/src/main/java/edu/oakland/sophomoreproject/configuration/SqliteConfiguration.java
@@ -1,4 +1,4 @@
-package configuration;
+package edu.oakland.sophomoreproject.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
