+++ bookmarket-backend/src/main/java/edu/oakland/sophomoreproject/controllers/ListingsController.java
@@ -1,0 +1,78 @@
+package edu.oakland.sophomoreproject.controllers;
+
+import edu.oakland.sophomoreproject.controllers.requests.CreateListingRequest;
+import edu.oakland.sophomoreproject.controllers.requests.UpdateListingRequest;
+import edu.oakland.sophomoreproject.controllers.responses.GetAllListingsResponse;
+import edu.oakland.sophomoreproject.controllers.responses.GetListingByIdResponse;
+import edu.oakland.sophomoreproject.authorization.SessionAuthorizer;
+import edu.oakland.sophomoreproject.components.ControllerUtils;
+import edu.oakland.sophomoreproject.dependencies.sqlite.sessions.Session;
+import jakarta.servlet.http.HttpServletRequest;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.sql.SQLException;
+
+@RestController
+public class ListingsController {
+	private final ControllerUtils controllerUtils;
+	private final SessionAuthorizer sessionAuthorizer;
+
+	public ListingsController(
+			ControllerUtils controllerUtils,
+			SessionAuthorizer sessionAuthorizer
+	) {
+		this.controllerUtils = controllerUtils;
+		this.sessionAuthorizer = sessionAuthorizer;
+	}
+
+	@GetMapping("/api/listing/{listingId}")
+	public ResponseEntity<GetListingByIdResponse> getListingById(
+			HttpServletRequest request,
+			@PathVariable("listingId") String listingId
+	) {
+		// ... do logic here
+
+		return ResponseEntity.ok().build();
+	}
+
+	@GetMapping("/api/listing")
+	public ResponseEntity<GetAllListingsResponse> getAllListings(HttpServletRequest request) {
+		// ... do logic here
+
+		return ResponseEntity.ok().build();
+	}
+
+	@PostMapping("/api/listing")
+	public ResponseEntity<Void> createListing(
+			HttpServletRequest request,
+			@RequestBody CreateListingRequest payload
+	) throws SQLException {
+		Session session = sessionAuthorizer.authorize(request);
+
+		// ... do logic here
+
+		HttpHeaders headers = controllerUtils.getHeadersWithSessionCookie(session);
+		return ResponseEntity.ok().headers(headers).build();
+	}
+
+	@PutMapping("/api/listing/{listingId}")
+	public ResponseEntity<Void> updateListing(
+			HttpServletRequest request,
+			@PathVariable String listingId,
+			@RequestBody UpdateListingRequest payload
+	) throws SQLException {
+		Session session = sessionAuthorizer.authorize(request);
+
+		// ... do logic here
+
+		HttpHeaders headers = controllerUtils.getHeadersWithSessionCookie(session);
+		return ResponseEntity.ok().headers(headers).build();
+	}
+}
