+++ bookmarket-backend/src/main/java/edu/oakland/sophomoreproject/controllers/requests/LoginRequest.java
@@ -1,4 +1,4 @@
-package controllers.requests;
+package edu.oakland.sophomoreproject.controllers.requests;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
