+++ bookmarket-backend/src/main/java/edu/oakland/sophomoreproject/controllers/responses/GetListingByIdResponse.java
@@ -1,4 +1,4 @@
-package controllers.responses;
+package edu.oakland.sophomoreproject.controllers.responses;
 
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
