+++ bookmarket-backend/src/main/java/edu/oakland/sophomoreproject/dependencies/sqlite/TableAccessor.java
@@ -1,4 +1,4 @@
-package dependencies.sqlite;
+package edu.oakland.sophomoreproject.dependencies.sqlite;
 
 import org.sqlite.SQLiteConfig;
